back/requirements.txt
fastapi
uvicorn
requests
sseclient-py
urllib3
pandas
numpy
scikit-learn
xgboost
joblib
sqlalchemy
psycopg2-binary
python-dotenv